Fix: Editing nodes with XNodeAttributes might lead to data corruption
If node was saved via the save button, the blur event handler was not
unbound and was called on the next blur event.

If this blur event resulted from clicking the edit
button on another node, the XNodeAttributes from that
node were set to the old one handled by unbound event handler,
because those attributes were always taken from the currently
edited node.