Prevent a race condition between "Replacing text" label and the save dialog, probably fixes #3484
Finishing the "Replacing text" label fade out after the save dialog
(or any other dialog) is made visible causes the BlockUI plugin to erase
its internal data necessary for restoring the dialog's html markup
to its original position and, as a result, to remove it from DOM tree
when dialog gets hidden. This makes it impossible to show this
dialog again.

Such a case could occur when text operation executed in the source editor
took a little bit longer.

This change fixes that by disabling the save button until the label
fades out.